fix: max height calculation in entities table
