+++ apps/hash-frontend/src/pages/entities.page.tsx
@@ -3,13 +3,7 @@ import { AsteriskRegularIcon } from "@hashintel/design-system";
 import { types } from "@local/hash-isomorphic-utils/ontology-types";
 import { isBaseUrl, OwnedById } from "@local/hash-subgraph";
 import { extractBaseUrl } from "@local/hash-subgraph/type-system-patch";
-import {
-  Box,
-  buttonClasses,
-  Container,
-  Typography,
-  useTheme,
-} from "@mui/material";
+import { Box, buttonClasses, Container, Typography } from "@mui/material";
 import { useRouter } from "next/router";
 import { NextSeo } from "next-seo";
 import { useCallback, useContext, useMemo } from "react";
@@ -25,16 +19,11 @@ import { FileCirclePlusRegularIcon } from "../shared/icons/file-circle-plus-regu
 import { FilesLightIcon } from "../shared/icons/files-light-icon";
 import { FilesRegularIcon } from "../shared/icons/files-regular-icon";
 import { getLayoutWithSidebar, NextPageWithLayout } from "../shared/layout";
-import { HEADER_HEIGHT } from "../shared/layout/layout-with-header/page-header";
-import { tableHeaderHeight } from "../shared/table-header";
 import { Button } from "../shared/ui";
 import { TabLink } from "../shared/ui/tab-link";
 import { Tabs } from "../shared/ui/tabs";
 import { EntitiesTable } from "./shared/entities-table";
-import {
-  TOP_CONTEXT_BAR_HEIGHT,
-  TopContextBar,
-} from "./shared/top-context-bar";
+import { TopContextBar } from "./shared/top-context-bar";
 import { WorkspaceContext } from "./shared/workspace-context";
 
 const contentMaxWidth = 1000;
@@ -107,8 +96,6 @@ const EntitiesPage: NextPageWithLayout = () => {
       ? `${entityType.schema.title} v${extractVersion(entityTypeId)}`
       : `${entityType.schema.title}s`
     : "Entities";
-
-  const theme = useTheme();
 
   const { entities, loading } = entityTypeEntitiesValue;
 
@@ -203,16 +190,6 @@ const EntitiesPage: NextPageWithLayout = () => {
           <EntitiesTable
             hideEntityTypeVersionColumn={!!entityTypeId}
             hidePropertiesColumns
-            height={
-              entities && entities.length > 10
-                ? `calc(100vh - (${
-                    HEADER_HEIGHT +
-                    TOP_CONTEXT_BAR_HEIGHT +
-                    179 +
-                    tableHeaderHeight
-                  }px + ${theme.spacing(5)} + ${theme.spacing(5)}))`
-                : undefined
-            }
           />
         </EntityTypeEntitiesContext.Provider>
       </Container>
